Removed observing windows loader
