+++ app/dataloader/__init__.py
@@ -1,5 +1,4 @@
 from .block_loader import BlockLoader  # noqa F401
 from .investigator_loader import InvestigatorLoader  # noqa F401
 from .observation_loader import ObservationLoader  # noqa F401
-from .observing_window_loader import ObservingWindowLoader  # noqa F401
 from .proposal_loader import ProposalLoader  # noqa F401
